health change and log variable working)
health change takes in an integer for health, we could maybe make one node for logging current health and another for logging the change in health. log variable has a lot of different options right now, but I think I can change it to be one node that can take whatever variable you give it. need to figure out if arrays break it or not